Improved error checking in Satellite class

+++ Hardware/SatelliteSample.cs
@@ -1,14 +1,11 @@
 ﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace AWS.Hardware
 {
     internal class SatelliteSample
     {
         [JsonProperty("windSpeed")]
-        public double? WindSpeed { get; set; } = null;
+        public int? WindSpeed { get; set; } = null;
 
         [JsonProperty("windDirection")]
         public int? WindDirection { get; set; } = null;
